refactor(workersvc): move sweeper (Sweep/resumePoolWaitRuns/publishSwept) to sweep.go
Pure same-package code motion (PRD #921 M1, seam 2). Moves the three sweeper
methods out of service.go into the new sweep.go, byte-identically (Sweep's
inline //nolint:gosec travels verbatim). hasLivePoller and stopKindFor stay
in service.go by design: hasLivePoller is a shared poller-liveness helper
Sweep does not call, and stopKindFor's callers all move to submit.go later.
No exported-API change, no behavior change; go doc output unchanged.
